AccountsWalletFragment CRUD basic, categories are missing

+++ app/src/main/java/com/camilo/wallet29/local_data/entity/AccountWalletEntity.kt
@@ -11,7 +11,7 @@ data class AccountWalletEntity(
     val id: Int,
     var name: String,
     var type: String,
-    var balance: Int,
+    var balance: Double,
     var description: String? = null,
     var color: Int,
     var icon: Int,
